De-duplicate mysql & postgres TLS code
