+++ sqlx-core/src/mysql/connection/tls.rs
@@ -1,8 +1,3 @@
-use sqlx_rt::{
-    fs,
-    native_tls::{Certificate, TlsConnector},
-};
-
 use crate::error::Error;
 use crate::mysql::connection::MySqlStream;
 use crate::mysql::protocol::connect::SslRequest;
@@ -46,34 +41,20 @@ async fn upgrade(stream: &mut MySqlStream, options: &MySqlConnectOptions) -> Res
 
     stream.flush().await?;
 
-    // FIXME: de-duplicate with postgres/connection/tls.rs
-
     let accept_invalid_certs = !matches!(
         options.ssl_mode,
         MySqlSslMode::VerifyCa | MySqlSslMode::VerifyIdentity
     );
+    let accept_invalid_host_names = !matches!(options.ssl_mode, MySqlSslMode::VerifyIdentity);
 
-    let mut builder = TlsConnector::builder();
-    builder
-        .danger_accept_invalid_certs(accept_invalid_certs)
-        .danger_accept_invalid_hostnames(!matches!(options.ssl_mode, MySqlSslMode::VerifyIdentity));
-
-    if !accept_invalid_certs {
-        if let Some(ca) = &options.ssl_ca {
-            let data = fs::read(ca).await?;
-            let cert = Certificate::from_pem(&data).map_err(Error::tls)?;
-
-            builder.add_root_certificate(cert);
-        }
-    }
-
-    #[cfg(not(feature = "_rt-async-std"))]
-    let connector = builder.build().map_err(Error::tls)?;
-
-    #[cfg(feature = "_rt-async-std")]
-    let connector = builder;
-
-    stream.upgrade(&options.host, connector.into()).await?;
+    stream
+        .upgrade(
+            &options.host,
+            accept_invalid_certs,
+            accept_invalid_host_names,
+            options.ssl_ca.as_deref(),
+        )
+        .await?;
 
     Ok(true)
 }
